requirements.txt
keystoneauth1
python-novaclient<8.0.0
python-glanceclient
python-neutronclient
